requirements.txt
streamlit
requests
selenium
webdriver-manager
python-dotenv
google-generativeai
pandas